Fix Android Stock < 4.4.4 <video> videoWidth/videoHeight bug
On Android Stock Browser > 4 and < 4.4.4 the 'loadedmetadata' event
fires without knowing the videoWidth and videoHeight values. This
causes this polyfill to fail on these browsers.

This fix adds 'startWork()' that monitors 'videoWidth' and
'videoHeight' until the earliest opportunity that they are available
and then calls 'doWork()' as normal.

This commit also fixes an issue in Android Stock Browsers > 4 and
< 4.4.4 where the first 'videoWidth' and 'videoHeight' dimensions
may be incorrect. We monitor the dimensions and if they change
during playback we run `doWork()` again.